fixed codesmells, adding new helper additionalPrice
